+++ src/grid_lbm/reset_grid_data.cpp
@@ -30,36 +30,36 @@ namespace hipoLBM
 
       inline void execute () override final
       {
-	constexpr Area L = Area::Local;
-	constexpr Traversal Tr = Traversal::All;
-	grid_data_lbm<Q>& grid_data = *GridDataQ;
-	domain_lbm<Q>& domain = *DomainQ;
-	grid<3>& Grid = domain.m_grid;
-	box<3>& Box = domain.m_box;
+        constexpr Area L = Area::Local;
+        constexpr Traversal Tr = Traversal::All;
+        grid_data_lbm<Q>& grid_data = *GridDataQ;
+        domain_lbm<Q>& domain = *DomainQ;
+        grid<3>& Grid = domain.m_grid;
+        box<3>& Box = domain.m_box;
 
-	// compute sizes
-	constexpr int Un = 5;
-	auto bx = Grid.build_box<L, Tr>();
-	int size_XYU = bx.get_length(0) * bx.get_length(1) * Un;
-	int size_YZU = bx.get_length(1) * bx.get_length(2) * Un;
-	int size_XZU = bx.get_length(0) * bx.get_length(2) * Un;
-	const int np = Box.number_of_points();
+        // compute sizes
+        constexpr int Un = 5;
+        auto bx = Grid.build_box<L, Tr>();
+        int size_XYU = bx.get_length(0) * bx.get_length(1) * Un;
+        int size_YZU = bx.get_length(1) * bx.get_length(2) * Un;
+        int size_XZU = bx.get_length(0) * bx.get_length(2) * Un;
+        const size_t np = Box.number_of_points();
 
-         grid_data.grid_size = np;
+        grid_data.grid_size = np;
 
-	if(grid_data.obst.size() != np)
-	{
-	  grid_data.f.resize(np*Q);
-	  grid_data.obst.resize(np);
-	  grid_data.m0.resize(np);
-	  grid_data.m1.resize(np);
-	  grid_data.fi_x_0.resize(size_YZU);
-	  grid_data.fi_x_l.resize(size_YZU);
-	  grid_data.fi_y_0.resize(size_XZU);
-	  grid_data.fi_y_l.resize(size_XZU);
-	  grid_data.fi_z_0.resize(size_XYU);
-	  grid_data.fi_z_l.resize(size_XYU);
-	}
+        if(grid_data.obst.size() != np)
+        {
+          grid_data.f.resize(np*Q);
+          grid_data.obst.resize(np);
+          grid_data.m0.resize(np);
+          grid_data.m1.resize(np);
+          grid_data.fi_x_0.resize(size_YZU);
+          grid_data.fi_x_l.resize(size_YZU);
+          grid_data.fi_y_0.resize(size_XZU);
+          grid_data.fi_y_l.resize(size_XZU);
+          grid_data.fi_z_0.resize(size_XYU);
+          grid_data.fi_z_l.resize(size_XYU);
+        }
       }
   };
